Extract payment handling into a separate file
